Backend changes, front end start

+++ Schiffskampfprofitrechner/Schiffskampfprofitrechner/shipFileGen.py
@@ -1,30 +1,30 @@
 # Ship File Generator
 # This tool will automatically read through a provided textfile and, after stripping whitespaces and dashes, create a .cs Class file using that name.
-
-import os, shutil
+# REDUNDANT
+import os
 
 def clearFolder(folder):
     for fil in os.listdir(folder): 
         abs_path = os.path.join(folder, fil)
         try:
             # If current file exists and is a .cs file, we clear it.
-            if (os.path.isfile(abs_path) and abs_path.endswith(".cs")):
+            if (os.path.isfile(abs_path) and abs_path.endswith(".cs") and not abs_path.endswith("Ship.cs")):
                 os.remove(abs_path)
         except Exception as e:
             # This exception will be raised if file is in use.
                 print(e)
 
-def classFile(fileName, name):
+def classFile(fileName, className, shipName, shipRes, shipFaction):
     class_struct = [
         "using System;\n",
         "\n",
         "namespace SKPR.Ships\n",
         "{\n",
-        "\tpublic class " + name + "\n",
+        "\tpublic class " + className + " " + ":" + " " + "Ship\n",
         "\t{\n",
-        "\t\tstring Name { get; set; }\n",
-        "\t\tint[] Resources { get; set; }\n",
-        "\t\tstring Faction { get; set; }\n",
+        "\t\tstring Name = " + '"' + shipName + '"' + ";\n",
+        "\t\tint[] Resources = " + shipRes + ";\n",
+        "\t\tstring Faction = " + '"' + shipFaction + '"' + ";\n",
         "\t}\n",
         "}"
         ]
@@ -39,5 +39,9 @@ clearFolder(os.getcwd())
 # .. and create a new .cs file
 with open("ShipReadouts.txt", "r") as file:
     for line in file:
-        short = line.replace('-','').replace(' ','').strip()
-        classFile(short + ".cs", short)
+        splitLine = line.split('.')
+        className = splitLine[0].replace('-','').replace(' ','')
+        shipName = splitLine[0]
+        shipRes = splitLine[1]
+        shipFaction = splitLine[2].replace('\n','')
+        classFile(className + ".cs", className, shipName, shipRes, shipFaction)
